Swipe and Take It Back Snackbar

+++ waypointsdk/src/main/java/com/neonapps/waypointsdk/Models/Review.kt
@@ -1,7 +1,7 @@
 package com.neonapps.waypointsdk.Models
 
 data class Review(
-    var rate : Int,
+    var rate : Double,
     var comment : String
 ) {
 }
